Move help widgets.
